# docs/swagger.yaml
definitions:
  payload.CreateSubscriptionPayload:
    properties:
      end_date:
        example: 08-2025
        type: string
      price:
        example: 400
        type: integer
      service:
        example: Yandex Plus
        type: string
      start_date:
        example: 07-2025
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    required:
    - price
    - service
    - start_date
    - user_id
    type: object
  payload.UpdateSubscriptionPayload:
    properties:
      end_date:
        example: 08-2025
        type: string
      price:
        example: 400
        type: integer
      service:
        example: Yandex Plus
        type: string
      start_date:
        example: 07-2025
        type: string
    required:
    - price
    - service
    - start_date
    type: object
  response.DeleteSubscriptionResponse:
    properties:
      message:
        example: subscription deleted successfully
        type: string
    type: object
  response.Error400:
    properties:
      code:
        example: "400"
        type: string
      error:
        example: invalid request body
        type: string
      request_id:
        example: req-12345
        type: string
    type: object
  response.Error404:
    properties:
      code:
        example: "404"
        type: string
      error:
        example: subscription not found
        type: string
      request_id:
        example: req-99999
        type: string
    type: object
  response.Error409:
    properties:
      code:
        example: "409"
        type: string
      error:
        example: subscription already exists
        type: string
      request_id:
        example: req-45678
        type: string
    type: object
  response.Error500:
    properties:
      code:
        example: "500"
        type: string
      error:
        example: internal server error
        type: string
      request_id:
        example: req-78901
        type: string
    type: object
  response.GetSubscriptionResponse:
    properties:
      created_at:
        example: "2025-07-01 12:00:00"
        type: string
      end_date:
        example: 08-2025
        type: string
      id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
      price:
        example: 400
        type: integer
      service:
        example: Yandex Plus
        type: string
      start_date:
        example: 07-2025
        type: string
      updated_at:
        example: "2025-07-01 12:00:00"
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    type: object
  response.ListSubscriptionsResponse:
    properties:
      count:
        example: 2
        type: integer
      subscriptions:
        items:
          $ref: '#/definitions/response.GetSubscriptionResponse'
        type: array
    type: object
  response.SubscriptionSummaryResponse:
    properties:
      total:
        example: 1200
        type: integer
    type: object
  response.SubscriptionUpdated:
    properties:
      message:
        example: subscription updated successfully
        type: string
    type: object
info:
  contact: {}
paths:
  /subscriptions:
    get:
      description: Возвращает список подписок с пагинацией
      parameters:
      - default: 10
        description: Лимит записей
        in: query
        name: limit
        type: integer
      - default: 0
        description: Смещение
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список подписок
          schema:
            $ref: '#/definitions/response.ListSubscriptionsResponse'
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/response.Error400'
        "404":
          description: Подписки не найдены
          schema:
            $ref: '#/definitions/response.Error404'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.Error500'
      summary: Получить список подписок
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Создаёт новую запись о подписке для пользователя
      parameters:
      - description: Новая подписка
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/payload.CreateSubscriptionPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.GetSubscriptionResponse'
        "400":
          description: Неверный JSON или валидация не пройдена
          schema:
            $ref: '#/definitions/response.Error400'
        "409":
          description: Подписка уже существует
          schema:
            $ref: '#/definitions/response.Error409'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.Error500'
      summary: Создать подписку
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Удаляет подписку по UUID
      parameters:
      - description: ID подписки (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Подписка успешно удалена
          schema:
            $ref: '#/definitions/response.DeleteSubscriptionResponse'
        "400":
          description: Неверный UUID
          schema:
            $ref: '#/definitions/response.Error400'
        "404":
          description: Подписка не найдена
          schema:
            $ref: '#/definitions/response.Error404'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.Error500'
      summary: Удалить подписку
      tags:
      - subscriptions
    get:
      description: Возвращает данные подписки по её UUID
      parameters:
      - description: ID подписки (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Подписка найдена
          schema:
            $ref: '#/definitions/response.GetSubscriptionResponse'
        "400":
          description: Неверный UUID
          schema:
            $ref: '#/definitions/response.Error400'
        "404":
          description: Подписка не найдена
          schema:
            $ref: '#/definitions/response.Error404'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.Error500'
      summary: Получить подписку по ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Обновляет существующую запись о подписке по ID
      parameters:
      - description: ID подписки (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/payload.UpdateSubscriptionPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Подписка успешно обновлена
          schema:
            $ref: '#/definitions/response.SubscriptionUpdated'
        "400":
          description: Неверный UUID или JSON
          schema:
            $ref: '#/definitions/response.Error400'
        "404":
          description: Подписка не найдена
          schema:
            $ref: '#/definitions/response.Error404'
        "409":
          description: Подписка уже существует
          schema:
            $ref: '#/definitions/response.Error409'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.Error500'
      summary: Обновить подписку
      tags:
      - subscriptions
  /subscriptions/summary:
    get:
      description: Подсчитывает общую стоимость подписок за указанный период с фильтрацией
        по user_id и названию сервиса
      parameters:
      - description: UUID пользователя
        example: '"60601fee-2bf1-4721-ae6f-7636e79a0cba"'
        in: query
        name: user_id
        type: string
      - description: Название сервиса
        example: '"Yandex Plus"'
        in: query
        name: service
        type: string
      - description: Дата начала периода YYYY-MM-DD
        example: '"2025-07-01"'
        in: query
        name: start_date
        required: true
        type: string
      - description: Дата конца периода YYYY-MM-DD
        example: '"2025-08-31"'
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Сумма подписок
          schema:
            $ref: '#/definitions/response.SubscriptionSummaryResponse'
        "400":
          description: Неверный формат запроса или параметров
          schema:
            $ref: '#/definitions/response.Error400'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/response.Error500'
      summary: Подсчёт суммарной стоимости подписок
      tags:
      - subscriptions
swagger: "2.0"